requirements.txt
streamlit
pandas
numpy
sentence-transformers
rank-bm25
faiss-cpu
huggingface_hub